lib: os: cbprintf: fix conditional initializer being inconsistent
In 7d56d6a this conditional initializer was added with different
requirements as what is used for the definition of cbprintf_package_desc.
There was edge cases where the initializer was added when the field does
not exist which was caught by clang-tidy. This commit ensures it uses
the same conditions to add the padding initializer.

Signed-off-by: Hein Wessels <heinwessels93@gmail.com>